refactor: VendingMachine 폴더 이동
