+++ src/ts/VendingMachine.ts
